quando adicionando o mesmo item soma + 1 a quantidade

+++ angularOnSpring/src/main/webapp/modulos/cliente/livro/lista-livro-controller.js
@@ -7,18 +7,40 @@ angular.module("livraria.cliente").controller("LivroClienteController", function
 		})
 	}
 	
+	var checarItem = function(item, carrinho) {
+		
+	}
+	
 	
 	$scope.adicionarItemNoCarrinho = function(item) {
-		console.log("adicionando item: " + item)
+
 		var carrinho = JSON.parse(localStorage.getItem("carrinho"))
 		if (carrinho == null){
 			carrinho = []
 		}
-		carrinho.push(item)
-		localStorage.setItem("carrinho", JSON.stringify(carrinho))
+		
+		var existeItem = carrinho.findIndex(function(currentValue) {
+			return currentValue.item.id === item.id
+		})
 		
 		
-		$rootScope.numeroDeItensCarrinho = carrinho.length
+		if(existeItem == -1){
+			var itemProduto = {'item': item, 'quantidade': 1}
+			
+			carrinho.push(itemProduto)
+			localStorage.setItem("carrinho", JSON.stringify(carrinho))
+			
+			$rootScope.numeroDeItensCarrinho = carrinho.length
+		} else{
+			carrinho[existeItem].quantidade += 1
+			localStorage.setItem("carrinho", JSON.stringify(carrinho))
+			
+		}
+		
+		
+		
+		
+		
 		
 		
 	}
